fix (operation) : error when creating a versioned document with a config.url method
Problem:
The versionId was missing in the populateURL  api.collection/area.update({ ... }) call arguments
because of that the update was threated as "create a new version" and required fields was missing.

Solution:
Add the missing versionId arg.

Note:
Maybe url should not be stored inside main table or verison table, as all versions should have the same url possibly localised
So probably an index table with columns : ownerId, locale, url would be better