update (operations) : add a mode "required" | "always" for the setDefaultValues function.
Problem:
Default values where added even if the field was not required, so it was not possible to set a field with a defaultValue to null.

Solution:
Add a mode param so we can adjust on where is a create/update to add or not the defaultValue.
The old behaviour still works for a create operation where all default values are added.
For all other update stuff set only for required field.